Tentative d'envoie ACK !

+++ src/pilot.c
@@ -23,13 +23,13 @@ int main(int argc, const char * argv[])
     long microsec = ((unsigned long long)time.tv_sec * 1000000) + time.tv_usec;
     char fifoName[250];
     sprintf(fifoName, "%ld", microsec);
-    
+
     // Make pilot fifo
     char path[MAX_BUF] = SHARED_FILE_PATH;
     strcat(path, fifoName);
     mkfifo(path, 0666);
-    
-    
+
+
     // Talk to control-tower
     printf(" - Waiting for signal -\n\n");
     FILE * ct;
@@ -46,7 +46,7 @@ int main(int argc, const char * argv[])
         exit(EXIT_FAILURE);
     }
     fclose(ct);
-    
+
     printf("Waiting for ATIS..\n");
     // Listen on pilot fifo
     while (1)
@@ -58,16 +58,28 @@ int main(int argc, const char * argv[])
             com_mess * ct_mess = read_message(fifo);
             if (ct_mess->header == HEADER_ATIS)
             {
-                printf("ATIS received, DC%s taking off ! Over.\n", fifoName);
-                unlink(path);
-                fclose(fifo);
-                break;
+                if(ct_mess->size == strlen(ct_mess->message))
+                {
+                    printf("ATIS OK, DC%s taking off ! Over.\n", fifoName);
+                    unlink(path);
+                    fclose(fifo);
+                    break;
+                }
+
             }
-            fclose(fifo);
+
+            FILE* rs = fopen(SHARED_FILE_PATH FIFO_IN_NAME, "w");
+            printf("ATIS KO, please send again !");
+            com_mess * mess2resend = encode_message(HEADER_HI, fifoName);
+            send_message(mess2resend, ct);
+            sleep(1);
+            }
+
         }
+
     }
-    
-    
+
+
     return 0;
 }
